Rollbacking entity prompts (#58)
* Fixing the Entity prompts issue
