--- generators/client/files.js
@@ -1,7 +1,0 @@
-module.exports = {
-    writeFiles
-};
-
-function writeFiles() {
-    this.copy('_dummy.txt', 'dummy.txt');
-}
